requirements.txt
linkedin
celery==5.3.4
tweepy==4.14.0
django-storages==1.14
django-cors-headers==4.2.0
stripe==6.4.0
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.45
certifi==2023.7.22
cffi==1.15.1
cryptography==41.0.3
distlib==0.3.7
Django==4.1.3
django-rest-knox==4.2.0
django-quill-editor==0.1.40
djangorestframework==3.14.0
filelock
gunicorn==21.2.0
importlib-metadata==6.8.0
iniconfig==2.0.0
install==1.3.5
packaging==23.1
Pillow==9.2.0
pipenv==2023.9.8
pippip==0.0.10
platformdirs==3.10.0
pluggy==1.0.0
psycopg2==2.9.7
py==1.11.0
pycparser==2.21
pyparsing==3.1.1
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==0.21.0
pytz==2023.3
pyjwt
sendgrid==6.9.7
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
virtualenv==20.24.5
virtualenv-clone==0.5.7
whitenoise==6.5.0
zipp==3.16.2
